requirements.txt
nasdaq-data-link>=1.0.4
xai-sdk>=1.7.0
pandas>=2.0.0
jinja2>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0
